+++ src/gr/gr_demod_base.h
@@ -78,8 +78,8 @@ public slots:
     void set_rx_sensitivity(double value, std::string gain_stage="");
     void set_squelch(int value);
     void set_gain(float value);
-    void set_agc_attack(float value);
-    void set_agc_decay(float value);
+    void set_agc_attack(int value);
+    void set_agc_decay(int value);
     void set_ctcss(float value);
     void enable_gui_const(bool value);
     void enable_gui_fft(bool value);
